test: unify further mazerunner scenarios
