+++ mazerunner-scenarios/detekt-baseline.xml
@@ -3,7 +3,21 @@
   <Blacklist></Blacklist>
   <Whitelist>
     <ID>EmptyFunctionBlock:Scenario.kt$Scenario${}</ID>
+    <ID>MagicNumber:BugsnagInitScenario.kt$BugsnagInitScenario$25</ID>
+    <ID>MagicNumber:StartupCrashFlushScenario.kt$StartupCrashFlushScenario$6000</ID>
+    <ID>MagicNumber:TestHarnessHooks.kt$&lt;no name provided&gt;$500</ID>
+    <ID>MagicNumber:TrimmedStacktraceScenario.kt$TrimmedStacktraceScenario$100000</ID>
+    <ID>TooGenericExceptionThrown:CrashHandlerScenario.kt$CrashHandlerScenario$throw RuntimeException("CrashHandlerScenario")</ID>
+    <ID>TooGenericExceptionThrown:CustomClientErrorFlushScenario.kt$CustomClientErrorFlushScenario$throw RuntimeException("ReportCacheScenario")</ID>
+    <ID>TooGenericExceptionThrown:DisableAutoNotifyScenario.kt$DisableAutoNotifyScenario$throw RuntimeException("Should never appear")</ID>
+    <ID>TooGenericExceptionThrown:IgnoredExceptionScenario.kt$IgnoredExceptionScenario$throw RuntimeException("Should never appear")</ID>
+    <ID>TooGenericExceptionThrown:ReportCacheScenario.kt$ReportCacheScenario$throw RuntimeException("ReportCacheScenario")</ID>
+    <ID>TooGenericExceptionThrown:StartupCrashFlushScenario.kt$StartupCrashFlushScenario$throw RuntimeException("Regular crash")</ID>
+    <ID>TooGenericExceptionThrown:StartupCrashFlushScenario.kt$StartupCrashFlushScenario$throw RuntimeException("Startup crash")</ID>
     <ID>TooManyFunctions:Scenario.kt$Scenario : ActivityLifecycleCallbacks</ID>
+    <ID>WildcardImport:AsyncErrorConnectivityScenario.kt$import com.bugsnag.android.*</ID>
+    <ID>WildcardImport:AsyncErrorDoubleFlushScenario.kt$import com.bugsnag.android.*</ID>
+    <ID>WildcardImport:AsyncErrorLaunchScenario.kt$import com.bugsnag.android.*</ID>
     <ID>WildcardImport:Scenario.kt$import com.bugsnag.android.*</ID>
   </Whitelist>
 </SmellBaseline>
