+++ mazerunner-scenarios/src/main/java/com/bugsnag/android/mazerunner/SomeException.kt
