+++ mazerunner-scenarios/src/main/java/com/bugsnag/android/mazerunner/scenarios/AsyncErrorConnectivityScenario.kt
@@ -5,13 +5,14 @@ import com.bugsnag.android.*
 
 /**
  * Tests that only 1 request is sent in the case where stored reports are concurrently flushed,
- * in the case that a connectivity change occurs after launch.
+ * in the case that a connectivity change occurs before launch.
  */
-internal class AsyncErrorLaunchScenario(config: Configuration,
-                                        context: Context) : Scenario(config, context) {
+internal class AsyncErrorConnectivityScenario(config: Configuration,
+                                              context: Context) : Scenario(config, context) {
 
     init {
-        config.delivery = createSlowDelivery(config)
+        val delivery = createSlowDelivery()
+        config.delivery = delivery
         config.setAutoCaptureSessions(false)
     }
 
@@ -19,8 +20,8 @@ internal class AsyncErrorLaunchScenario(config: Configuration,
         super.run()
 
         writeErrorToStore(Bugsnag.getClient())
-        flushErrorStoreOnLaunch(Bugsnag.getClient())
         flushErrorStoreAsync(Bugsnag.getClient())
+        flushErrorStoreOnLaunch(Bugsnag.getClient())
     }
 
 }
