+++ mazerunner-scenarios/src/main/java/com/bugsnag/android/mazerunner/scenarios/AutoSessionScenario.kt
@@ -5,19 +5,16 @@ import android.content.Intent
 import com.bugsnag.android.Bugsnag
 import com.bugsnag.android.Configuration
 import com.bugsnag.android.flushAllSessions
-import com.bugsnag.android.mazerunner.SecondActivity
 
 /**
  * Sends a manual session payload to Bugsnag.
  */
 internal class AutoSessionScenario(config: Configuration,
                                    context: Context) : Scenario(config, context) {
+
     override fun run() {
         super.run()
-        config.setAutoCaptureSessions(true)
         Bugsnag.init(context, config)
-        Bugsnag.setUser("123", "user@example.com", "Joe Bloggs")
-        context.startActivity(Intent(context, SecondActivity::class.java))
     }
 
 }
