+++ mazerunner-scenarios/src/main/java/com/bugsnag/android/mazerunner/scenarios/BeforeSendScenario.java
@@ -1,25 +1,25 @@
 package com.bugsnag.android.mazerunner.scenarios;
 
-import com.bugsnag.android.Bugsnag;
 import com.bugsnag.android.BeforeSend;
+import com.bugsnag.android.Bugsnag;
 import com.bugsnag.android.Configuration;
 import com.bugsnag.android.Report;
-import com.bugsnag.android.Severity;
 
 import android.content.Context;
 import androidx.annotation.NonNull;
 
+public class BeforeSendScenario extends Scenario {
 
-public class NotifyBeforeSendScenario extends Scenario {
-
-    public NotifyBeforeSendScenario(@NonNull Configuration config, @NonNull Context context) {
+    /**
+     * Constructs a scenario which adds a beforeSend block
+     */
+    public BeforeSendScenario(@NonNull Configuration config, @NonNull Context context) {
         super(config, context);
         config.setAutoCaptureSessions(false);
         config.addBeforeSend(new BeforeSend() {
             @Override
             public boolean run(Report report) {
-                report.getError().setContext("RESET");
-                report.getError().setSeverity(Severity.ERROR);
+                report.getError().setContext("UNSET");
 
                 return true;
             }
@@ -29,6 +29,10 @@ public class NotifyBeforeSendScenario extends Scenario {
     @Override
     public void run() {
         super.run();
-        Bugsnag.notify(new Exception("Registration failure"));
+        String metadata = getEventMetaData();
+        if (metadata != null && metadata.equals("non-crashy")) {
+            return;
+        }
+        throw new RuntimeException("Ruh-roh");
     }
 }
