+++ mazerunner-scenarios/src/main/java/com/bugsnag/android/mazerunner/scenarios/InForegroundScenario.kt
