+++ mazerunner-scenarios/src/main/java/com/bugsnag/android/mazerunner/scenarios/NotifyBeforeSendUnsetContextScenario.java
@@ -1,7 +1,7 @@
 package com.bugsnag.android.mazerunner.scenarios;
 
-import com.bugsnag.android.Bugsnag;
 import com.bugsnag.android.BeforeSend;
+import com.bugsnag.android.Bugsnag;
 import com.bugsnag.android.Configuration;
 import com.bugsnag.android.Report;
 import com.bugsnag.android.Severity;
@@ -9,11 +9,16 @@ import com.bugsnag.android.Severity;
 import android.content.Context;
 import androidx.annotation.NonNull;
 
-
 public class NotifyBeforeSendUnsetContextScenario extends Scenario {
 
-    public NotifyBeforeSendUnsetContextScenario(@NonNull Configuration config, @NonNull Context context) {
+    /**
+     * Constructs a scenario where notify is called with a beforeSend block and the context is
+     * unset
+     */
+    public NotifyBeforeSendUnsetContextScenario(@NonNull Configuration config,
+                                                @NonNull Context context) {
         super(config, context);
+        config.setAutoCaptureSessions(false);
         config.addBeforeSend(new BeforeSend() {
             @Override
             public boolean run(Report report) {
