+++ mazerunner-scenarios/src/main/java/com/bugsnag/android/mazerunner/scenarios/Wait.kt
@@ -12,8 +12,8 @@ internal class Wait(config: Configuration, context: Context) : Scenario(config, 
         super.run()
         val thread = HandlerThread("HandlerThread")
         thread.start()
-        Handler(thread.looper).post {
+        Handler(thread.looper).post(Runnable {
             flushAllSessions()
-        }
+        })
     }
 }
