--- tests/features/fixtures/mazerunner/src/main/java/com/bugsnag/android/mazerunner/SomeException.kt
@@ -1,4 +1,0 @@
-package com.bugsnag.android.mazerunner
-
-internal class SomeException() : Exception() {
-}
